+++ raft/vote_test.go
@@ -9,13 +9,13 @@ func TestHandleVoteRequest(t *testing.T) {
 		req          Message
 		wantSuccess  bool
 		wantTerm     uint64
-		wantVotedFor uint64
+		wantVotedFor int
 		wantState    stateType
 	}{
 		{
 			name: "rejects stale term",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3}, withTerm(5))
+				return newTestCore(t, 1, []int{1, 2, 3}, withTerm(5))
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 2, Term: 3},
 			wantSuccess:  false,
@@ -26,7 +26,7 @@ func TestHandleVoteRequest(t *testing.T) {
 		{
 			name: "grants vote to up-to-date candidate with empty log",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3})
+				return newTestCore(t, 1, []int{1, 2, 3})
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 2, Term: 1},
 			wantSuccess:  true,
@@ -37,7 +37,7 @@ func TestHandleVoteRequest(t *testing.T) {
 		{
 			name: "steps down and votes when candidate's term is higher, even from a leader",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3}, withTerm(2), withState(LeaderState))
+				return newTestCore(t, 1, []int{1, 2, 3}, withTerm(2), withState(LeaderState))
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 2, Term: 5},
 			wantSuccess:  true,
@@ -48,7 +48,7 @@ func TestHandleVoteRequest(t *testing.T) {
 		{
 			name: "refuses a second vote to a different candidate in the same term",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3}, withTerm(3), withVotedFor(2))
+				return newTestCore(t, 1, []int{1, 2, 3}, withTerm(3), withVotedFor(2))
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 4, Term: 3},
 			wantSuccess:  false,
@@ -59,7 +59,7 @@ func TestHandleVoteRequest(t *testing.T) {
 		{
 			name: "re-grants to the same candidate already voted for in this term",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3}, withTerm(3), withVotedFor(2))
+				return newTestCore(t, 1, []int{1, 2, 3}, withTerm(3), withVotedFor(2))
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 2, Term: 3},
 			wantSuccess:  true,
@@ -70,7 +70,7 @@ func TestHandleVoteRequest(t *testing.T) {
 		{
 			name: "rejects a candidate whose log ends in an older term",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3}, withTerm(3), withLog(Entry{Term: 2}, Entry{Term: 3}))
+				return newTestCore(t, 1, []int{1, 2, 3}, withTerm(3), withLog(Entry{Term: 2}, Entry{Term: 3}))
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 2, Term: 3, LastLogIndex: 2, LastLogTerm: 2},
 			wantSuccess:  false,
@@ -81,7 +81,7 @@ func TestHandleVoteRequest(t *testing.T) {
 		{
 			name: "rejects a candidate with the same last term but a shorter log",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3}, withTerm(3), withLog(Entry{Term: 2}, Entry{Term: 3}))
+				return newTestCore(t, 1, []int{1, 2, 3}, withTerm(3), withLog(Entry{Term: 2}, Entry{Term: 3}))
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 2, Term: 3, LastLogIndex: 1, LastLogTerm: 3},
 			wantSuccess:  false,
@@ -92,7 +92,7 @@ func TestHandleVoteRequest(t *testing.T) {
 		{
 			name: "grants a candidate with the same last term and an equal-or-longer log",
 			core: func(t *testing.T) *Core {
-				return newTestCore(t, 1, []uint64{1, 2, 3}, withTerm(3), withLog(Entry{Term: 2}, Entry{Term: 3}))
+				return newTestCore(t, 1, []int{1, 2, 3}, withTerm(3), withLog(Entry{Term: 2}, Entry{Term: 3}))
 			},
 			req:          Message{Type: MsgVoteRequest, FromId: 2, Term: 3, LastLogIndex: 2, LastLogTerm: 3},
 			wantSuccess:  true,
@@ -131,13 +131,13 @@ func TestHandleVoteRequest(t *testing.T) {
 
 func TestHandleVoteResponse(t *testing.T) {
 	t.Run("majority of grants elects the candidate and initializes leader state", func(t *testing.T) {
-		c := newTestCore(t, 1, []uint64{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
+		c := newTestCore(t, 1, []int{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
 		c.Step(Message{Type: MsgVoteResponse, FromId: 2, Term: 1, Success: true})
 
 		if c.state != LeaderState {
 			t.Fatalf("state = %v, want LeaderState", c.state)
 		}
-		for _, peer := range []uint64{2, 3} {
+		for _, peer := range []int{2, 3} {
 			if got, want := c.nextIndex[peer], c.lastIndex()+1; got != want {
 				t.Fatalf("nextIndex[%d] = %d, want %d (Figure 2: reinitialized to leader's last log index + 1)", peer, got, want)
 			}
@@ -148,14 +148,14 @@ func TestHandleVoteResponse(t *testing.T) {
 	})
 
 	t.Run("a single vote (self only) is not a majority of 3", func(t *testing.T) {
-		c := newTestCore(t, 1, []uint64{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
+		c := newTestCore(t, 1, []int{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
 		if c.state != CandidateState {
 			t.Fatalf("state = %v, want CandidateState (self-vote alone must not win a 3-node election)", c.state)
 		}
 	})
 
 	t.Run("a denied vote is not counted", func(t *testing.T) {
-		c := newTestCore(t, 1, []uint64{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
+		c := newTestCore(t, 1, []int{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
 		c.Step(Message{Type: MsgVoteResponse, FromId: 2, Term: 1, Success: false})
 		if c.state != CandidateState {
 			t.Fatalf("state = %v, want CandidateState", c.state)
@@ -166,7 +166,7 @@ func TestHandleVoteResponse(t *testing.T) {
 	})
 
 	t.Run("a response from an election we've already abandoned (stale term) is ignored", func(t *testing.T) {
-		c := newTestCore(t, 1, []uint64{1, 2, 3}, withState(CandidateState), withTerm(6), withVotesGranted(1))
+		c := newTestCore(t, 1, []int{1, 2, 3}, withState(CandidateState), withTerm(6), withVotesGranted(1))
 		c.Step(Message{Type: MsgVoteResponse, FromId: 2, Term: 5, Success: true})
 		if c.state != CandidateState {
 			t.Fatalf("state = %v, want CandidateState (pitfall #6: stale-term reply must not be treated as current)", c.state)
@@ -177,7 +177,7 @@ func TestHandleVoteResponse(t *testing.T) {
 	})
 
 	t.Run("a higher-term response steps the candidate down", func(t *testing.T) {
-		c := newTestCore(t, 1, []uint64{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
+		c := newTestCore(t, 1, []int{1, 2, 3}, withState(CandidateState), withTerm(1), withVotesGranted(1))
 		c.Step(Message{Type: MsgVoteResponse, FromId: 2, Term: 5, Success: false})
 		if c.state != FollowerState {
 			t.Fatalf("state = %v, want FollowerState", c.state)
